HHH-16900 Move EntityMutationOperationGroup to the same package as related APIs
