+++ hibernate-core/src/main/java/org/hibernate/sql/model/EntityMutationOperationGroup.java
@@ -4,10 +4,9 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html.
  */
-package org.hibernate.engine.jdbc.mutation.internal;
+package org.hibernate.sql.model;
 
 import org.hibernate.persister.entity.mutation.EntityMutationTarget;
-import org.hibernate.sql.model.MutationOperationGroup;
 
 public interface EntityMutationOperationGroup extends MutationOperationGroup {
